keypoints/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
aiohappyeyeballs=2.4.3=pypi_0
aiohttp=3.10.10=pypi_0
aiosignal=1.3.1=pypi_0
albucore=0.0.17=pypi_0
albumentations=1.4.18=pypi_0
annotated-types=0.7.0=pypi_0
asttokens=2.4.1=pypi_0
attrs=24.2.0=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.9.24=h06a4308_0
certifi=2024.8.30=pypi_0
charset-normalizer=3.4.0=pypi_0
click=8.1.7=pypi_0
contourpy=1.3.0=pypi_0
cycler=0.12.1=pypi_0
decorator=5.1.1=pypi_0
docker-pycreds=0.4.0=pypi_0
efficientnet-pytorch=0.7.1=pypi_0
eval-type-backport=0.2.0=pypi_0
executing=2.1.0=pypi_0
expat=2.6.3=h6a678d5_0
filelock=3.16.1=pypi_0
fonttools=4.54.1=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2024.9.0=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
huggingface-hub=0.25.2=pypi_0
idna=3.10=pypi_0
imageio=2.36.0=pypi_0
intensity-normalization=2.2.4=pypi_0
ipdb=0.13.13=pypi_0
ipython=8.29.0=pypi_0
itk=5.4.0=pypi_0
itk-core=5.4.0=pypi_0
itk-filtering=5.4.0=pypi_0
itk-io=5.4.0=pypi_0
itk-numerics=5.4.0=pypi_0
itk-registration=5.4.0=pypi_0
itk-segmentation=5.4.0=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.4=pypi_0
joblib=1.4.2=pypi_0
kiwisolver=1.4.7=pypi_0
lazy-loader=0.4=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
lightning-utilities=0.11.8=pypi_0
markupsafe=3.0.1=pypi_0
matplotlib=3.9.2=pypi_0
matplotlib-inline=0.1.7=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.1.0=pypi_0
munch=4.0.0=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.4.1=pypi_0
nibabel=3.2.2=pypi_0
numpy=1.26.4=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=9.1.0.70=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.20.5=pypi_0
nvidia-nvjitlink-cu12=12.6.77=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
opencv-contrib-python=4.10.0.84=pypi_0
opencv-python=4.10.0.84=pypi_0
openssl=3.0.15=h5eee18b_0
packaging=24.1=pypi_0
pandas=2.2.3=pypi_0
parso=0.8.4=pypi_0
patsy=0.5.6=pypi_0
pexpect=4.9.0=pypi_0
pillow=11.0.0=pypi_0
pip=24.2=py312h06a4308_0
platformdirs=4.3.6=pypi_0
pretrainedmodels=0.7.4=pypi_0
prompt-toolkit=3.0.48=pypi_0
propcache=0.2.0=pypi_0
protobuf=5.28.2=pypi_0
psutil=6.0.0=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.3=pypi_0
py-cpuinfo=9.0.0=pypi_0
pydantic=2.9.2=pypi_0
pydantic-core=2.23.4=pypi_0
pydicom=2.4.4=pypi_0
pygments=2.18.0=pypi_0
pymedio=0.2.14=pypi_0
pyparsing=3.2.0=pypi_0
pystackreg=0.2.8=pypi_0
python=3.12.7=h5148396_0
python-dateutil=2.9.0.post0=pypi_0
pytorch-lightning=2.4.0=pypi_0
pytz=2024.2=pypi_0
pyyaml=6.0.2=pypi_0
readline=8.2=h5eee18b_0
requests=2.32.3=pypi_0
safetensors=0.4.5=pypi_0
scikit-fuzzy=0.5.0=pypi_0
scikit-image=0.24.0=pypi_0
scikit-learn=1.5.2=pypi_0
scipy=1.14.1=pypi_0
seaborn=0.13.2=pypi_0
segmentation-models-pytorch=0.3.4=pypi_0
sentry-sdk=2.17.0=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=75.1.0=py312h06a4308_0
shapely=2.0.6=pypi_0
simpleitk=2.4.0=pypi_0
six=1.16.0=pypi_0
smmap=5.0.1=pypi_0
sqlite=3.45.3=h5eee18b_0
stack-data=0.6.3=pypi_0
statsmodels=0.14.4=pypi_0
sympy=1.13.3=pypi_0
threadpoolctl=3.5.0=pypi_0
tifffile=2024.9.20=pypi_0
timm=0.9.7=pypi_0
tk=8.6.14=h39e8969_0
torch=2.4.1=pypi_0
torchmetrics=1.4.3=pypi_0
torchvision=0.19.1=pypi_0
tqdm=4.66.5=pypi_0
traitlets=5.14.3=pypi_0
triton=3.0.0=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2024.2=pypi_0
ultralytics=8.3.15=pypi_0
ultralytics-thop=2.0.9=pypi_0
urllib3=2.2.3=pypi_0
wandb=0.18.3=pypi_0
wcwidth=0.2.13=pypi_0
wheel=0.44.0=py312h06a4308_0
xz=5.4.6=h5eee18b_1
yarl=1.15.4=pypi_0
zlib=1.2.13=h5eee18b_1
